written most of dynamic shot controller. need to write drivetrain controller now but that's so much energy so pestolink update it is

+++ include/Arm.h
@@ -34,7 +34,7 @@ class Arm
             armSetAngle = angle;
         }
 
-        void setDynamic(double angle);
+        void setDynamic(float angle);
 
 
     private:
